Lucene.Net.Analysis.Nori.Support: Upgraded FST to Lucene 8.2.0 implementation

+++ src/Lucene.Net.Analysis.Nori/Support/Util/Fst/ForwardBytesReader.cs
@@ -50,9 +50,9 @@ namespace Lucene.Net.Support.Util.Fst
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public override void SkipBytes(int count)
+        public override void SkipBytes(long count)
         {
-            pos += count;
+            pos += (int)count;
         }
 
         public override long Position
